fixed zipkin-web to actually return values from servers
Ruby 1.9.3 fixed a problem where do loops were sometimes not returning their
last value. However, for compatibility with 1.9.2, the with_transport do loops
need their last value to be assigned to a variable with a greater scope than the
do loop, ie to be initialized outside of the do loop and then assigned within.
Either the necessary ruby version should be specified in the documentation, or
this fix needs to be pulled in.

Author: @mosesn
Fixes #68
URL: https://github.com/twitter/zipkin/pull/68