Improve local js,css development
Making changes to js,css can be painful with a packaged jar since a
compilation is needed to repackage any new changes. This change allows
for the paths to the static resources to be configured at runtime.
While developing locally, you can then set the prefix to bypass the jar
and the browser will resolve it on your local filesystem.

Author: @franklinhu
Fixes #112
URL: https://github.com/twitter/zipkin/pull/112