Update installation documentation
The current installation documentation didn't get me very far in getting set up,
in part because it seems outdated. I added a document with the commands I ran on
my MBP to get Zipkin working and cleaned up some of the explanation around it.
NOTE: this is not quite ready for merging as there are a few TODOs that need
review where I wasn't sure if some steps were still required.

The next step is
to update the website (in the gh-pages branch) to reflect these changes; or we
might decide that the master branch is the Source of Truth and we should just
point the website to /docs for installation instructions.

Author: @IceCreamYou
Fixes #251
URL: https://github.com/twitter/zipkin/pull/251