Added better support for Redis.
Fixed a bug in
zipkin-redis:com.twitter.zipkin.storage.redis.RedisSortedSetMap where a
Either[Future[ZRangeResults], Future[Seq[ChannelBuffer]]] was being returned instead of
a Future[ZRangeResults]. The method in question will always return a
ZRangeResult because it is getting a sorted map, so we simply get the
left item of the Either.

Added bin/collector-redis and bin/query-redis for easier
setup of the collector and querier with redis support.

Added zipkin-redis/config with the scribe configuration files for a
sample client and server. These are simply example files.