de-duplicate span records in anormdb
The AnormDB data model is a bit wrong. Spans don't get stored as unique records.
Because of this a span from the server and a span from the server get saved to
the span table as two different records. Reading them out then causes duplicate
data to be displayed in the UI.

Author: @sprsquish
Fixes #356
URL: https://github.com/twitter/zipkin/pull/356