finagle-core: Introduce Addr.Metadata into Addr.Bound
Problem:

There is no way for Namers to surface parameters to the client stack. Since
resolved clusters may be regionally diverse, etc; we'd like a way for clients'
endpoint stacks to use this information to instrument timeouts and stats.

Solution:

Introduce Addr.Metadata, a map on String to Any, that resolvers & namers may
use to surface parameters to client construction. This breaks the Addr.Bound
api, since it is a case class.

RB_ID=547110
TBR=true