Constrains Travis to working configuration and enables caching
The openjdk7 build doesn't work and #423 broke the one that worked. This
constrains Travis to what can pass, allowing contributors to make use of
it as opposed to be constantly distracted by a half-broke build. In the
future, we'll want to fixup the current build and re-add openjdk.

This also attempts to the build time by caching artifacts. This may help
in reducing flakiness.

Signed-off-by: Adrian Cole <acole@twitter.com>

RB_ID=706898
TBR=true