Ports specs/jmock tests to scalatest/mockito and cleanup
Before, tests were a mix of scalatest and specs, and a mix of jmock and
mockito. This migrates all tests to scalatest/mockito and removes the
`@RunWith` annotation, which isn't necessary in SBT or Intellij.

All tests are now `FunSuite` except `QueryServiceSpec`, which is vast
and better organized as `WordSpec` for now.

Almost all tests use Matchers, and the ones that don't aren't hurting
anything, so could later be moved to Matchers if desireable.

Finally, doing this simplifies the build: we now have a single test
dependency bundle, and it is far simpler than before.

Fixes #493