Refactors SBT plugin configuration
The SBT plugin configuration we used was out-of-date, and for example
caused us to download different versions of scrooge. We also had an
unused copy/pasted unidoc plugin.

This change clears out unused and/or confusing plugin configuration. It
also updates plugins we actually use to current versions.

The sbt-assembly plugin was updated to 0.11.2, as versions beyond that
need significant configuration work.