Moves zipkin-hbase to its own repository
zipkin-hbase integration tests regularly cause our builds to crash. This
has endangered the productivity of the project. Since this has never
been released, this burden isn't justifiable.

This change isolates hbase to its own repository, where it can either
be maintained, or be archived.

It is an acknowledged reality that this repository won't work until we
release zipkin artifacts it depends on, and perhaps perform other
modularity work.

https://github.com/openzipkin/zipkin-hbase