Add build config for executable jars
All the executable projects (the zipkin-*-services and zipkin-web)
now have a *-all.jar which is executable. Still to do: package the
config files for the collector service along with the code (at the
moment you have to have them locally and run with
"-f <path_to>/collector-<xxx>/scala".