Redis storage rewrite
Redis contained a lot of normalization, for example substrate types such
as RedisHash and RedisListMap. These seemed like a good idea, but the
indirection makes the code extremely difficult to follow. This was even
harder considering use of implicits and the amount of dead code.

This pull requests removes all use of implicits, complex expressions,
and reduces the call stack depth by inlining code that was formerly used
once. This also removes all dead code and backfills some tests.