Makes SpanStore and Aggregates implementable in Java
This makes the critical storage interfaces, SpanStore and Aggregates,
abstract classes, and adds implementations in the java test tree to
prove they can compile with Java 7.

SpanStore was a mixture of traits, which made it impossible to write in
Java. Moreover, eventhough all implementations close synchronously, it
mixed in Twitter's Closeable, which was both a trait and also async.

Aggregates was a trait and implemented a synchronous close method
already.

See #451
Closes #584