Removes unused canned scripts and tar.gz publication
We shouldn't be storing things not used, or complicating publication
with duplicate artifacts.

The scripts stored in zipkin's repo are not used at Twitter, not used by
the dist routine, not used by our bin directory, and not used in docker.

At the moment, the tar.gz dist isn't used by anything either. When our
artifact publication works, there's nothing in the tar.gz we can do that
we can't do with the fat jar. Once we have versioned fat jars, we should
change docker to use them.. we don't need a tar.gz IOTW.