Consolidates and cleans up getTraces thrift query
This deprecates thrift calls for zipkin-web data conversion and
consolidates usage to a getTraces query. We will drop the obsolete
methods in 1.5.0, resulting in a much cleaner thrift api.

Before this change, there was a leaky abstraction between the query and
the web packages. For example, types like TraceCombo were introduced
eventhough they were only a single stage of data conversion and only
used to support one web view. The web component now handles data
conversions directly. Moreover, the web component duplicated work by
doing an id scan which implied looking up spans before looking up spans.

This also cleans up a couple questionable api concerns: adjuster and
binary annotation query.

In practice, there was only one type of adjustment, time-skew, and this
was almost always enabled. The code was confused by an idea that there
would be many types of adjustments, and in some places time-skew was
applied multiple times. This moves it to a boolean flag, just like the
http api.

The binary annotation query in thrift was a weird one. It allows the
user to send a string -> string query only. However, the signature
accepted literally the BinaryAnnotation type, which has other members,
and might not be a string. This clarifies the concern by only accepting
a map of string -> string.