Integrates Kafka receiver into the zipkin-collector-service process
This starts integration with kafka by triggering it on presence of the
`KAFKA_ZOOKEEPER` environment variable. Refinements to defaults will
occur in later change, with a goal to have the minimum viable setup and
one that is easy to configure with docker.

For example, if you are starting from a zipkin's source tree, you might
assign this way:

```bash
$ KAFKA_ZOOKEEPER=127.0.0.1:2181 bin/collector
$ KAFKA_ZOOKEEPER=127.0.0.1:2181 KAFKA_TOPIC=notzipkin bin/collector
```

Note: Enabling Kafka doesn't disable the Scribe listener as it's part of
the Collector's thrift api alongside aggregate dependencies. Future work
might decouple that relationshop.

See #682
Closes #547