Drops support for annotation.duration
This drops annotation.duration, an optional field rarely used in zipkin.
It causes more problems than it solves.

Implementations handled `annotation.duration` in many ways. Some advised
against using it (circonus), others exposed hooks not used (brave), some
stored the result of timestamp subtraction (spring-cloud-sleuth), or put
constants like 1 (htrace), yet others stored bugs (wrong timeunit in
anormdb). Many didn't support it at all, like ruby and python tracers.

Internally, zipkin has a lot of logic around annotation durations, but
it does timestamp math, in other words, it ignores this value.

On the front end, `Annotation.duration` has space dedicated to it in
the UI, which is often blank. It is not queryable and actually was used
to set an incorrect width.

I could go on, but bottom-line is that this doesn't add value to the
zipkin experience, rather it subtracts from it. Dropping this field is
safe as it was an optional thrift field to begin with, and it wasn't
useful as implemented.

https://github.com/twitter/finagle/blob/develop/finagle-core/src/main/scala/com/twitter/finagle/filter/JvmFilter.scala#L23
https://github.com/apache/incubator-htrace/blob/master/htrace-zipkin/src/main/java/org/apache/htrace/zipkin/HTraceToZipkinConverter.java#L190
https://github.com/circonus-labs/libmtev/blob/master/src/utils/mtev_zipkin.h#L215
https://github.com/wadey/go-zipkin/blob/master/trace.go#L144
https://github.com/spring-cloud/spring-cloud-sleuth/blob/master/spring-cloud-sleuth-zipkin/src/main/java/org/springframework/cloud/sleuth/zipkin/ZipkinSpanListener.java#L205
https://github.com/openzipkin/brave/blob/master/brave-core/src/main/java/com/github/kristofa/brave/AnnotationSubmitter.java#L42

The only use of `Annotation.duration` I could find that wasn't derived
from other span data was finagle's gc annotation. Finagle's GC is a neat
annotation, but better stored as a binary one, instead of a semi-
structured string like this:

```
value=Gc(10,0.PSScavenge,2015-09-17 12:57:12 +0000,525.milliseconds+970.microseconds)
duration=525970
```

Note that even when we remove this, the duration is still present in the
string, in millis.

For example, if it were a binary annotation, you could store the entire
GC object as a json, and use the key `jvm.gc` instead. This is easier
than parsing the toString of a scala object.