Adds MySQL backend
Adds an experimental MySQL backend which uses a simpler table structure
than upstream. It is simpler as it keeps a single table for annotations
regardless of whether they are "binary" or not.

While not yet hooked up, this passes all tests in SpanStoreSpec, and I
also succesfully deployed it to EC2, targetting an RDS MySQL instance.