Converts the communication between web and query to http
Formerly, Zipkin's UI communicated with the query server via thrift RPC.
This means that any implementation that supports the UI needs to run a
Thrift listener. There's many problems with this.

* Thrift RPC isn't supported in most frameworks, hence a learning curve
* Thrift RPC isn't supported by loadbalancers, so you have to go TCP
* The only thrift RPC tracing implementation is custom to finagle
* Finagle has their own TCP multiplexer, which isn't supported elsewhere

Worst than all of this is that the choice is a poor match to power a UI.
zipkin-web literally converts these calls to http+json, and there's no
other consumer of the query api. Moving this link to normal HTTP/json
will make Zipkin easier to maintain, interop with and troubleshoot.