Reverts code/config that dodges java.beans
We were dodging access to the `java.beans` package to allow zipkin to
run on JRE compact3 profile, as the OS image becomes quite small. This
has become an unsustainable goal as too many libraries, now also Finatra
aren't designed for this. This rolls back things to not avoid
java.beans.