Removes implicit type of numeric binary annotations
The only types in json we can be sure about are strings and booleans. In
designing the shape of json, I over-optimized, dropping type keys on
numerics, depending on jackson defaults. This was dangerous as subtle
changes to config will end up corrupting thrifts. This change requires
numbers to be explicitly typed in the json format, ex. with "I32" if it
should be an integer.