Settles on startTs/endTs and fixes Dependencies UI
Troubleshooting subtle problems was unnecessarily difficult as we used
many ways to refer to the start and end of a span in epoch microseconds.

This settles on `startTs` and `endTs`, which are the dominant names. By
doing so, it is cognitively easier as, for example, we can see that
`Span.endTs` relates directly to the query parameter `endTs`, as well as
the upper bound of `Dependencies.endTs`.

This bumps the version because it fixes the dependencies api endpoint to
use normal query parameters as opposed to path expressions for `startTs`
and `endTs`. This also renames the thrift fields for dependencies,
although that isn't breaking as names aren't encoded in thrift
serialization.

Most importantly, this fixes the Dependencies UI, which has been broken
for a while.