Adds Server Receive and Server Send annotations
This adds a Brave tracer, adding Server Receive and Server Send
annotations for `zipkin-query`.

MySQL tracing is possible, but needs revision to spring boot
configuration. We need brave-mysql in the same classloader as HikariCP,
which shading would accomplish, but I was unable to get a working boot
configuration to do that.