Simplified getDependencies and reworks tests to allow on-demand
Before, `DependencyStore.getDependencies()` returned an envelope type
which was always unwrapped into links. This simplifies that by just
returning the links.

This also adds an in-memory parser which converts spans into dependency
links. This is needed to do tests on backends that don't compute links
on-demand. `DependencyStore.storeDependencies()` should be optional as
backends that aggregate on-demand don't need the intermediate form.