Updates to scala zipkin 1.17 and removes scribe
This updates to the latest scala zipkin 1.17, which has a compatible
schema with `zipkin-java-jdbc`.

This deletes the scribe code, as it is no longer needed. Those needing
scribe can run the scala collector against the same database like so:

```yaml
collector:
  image: openzipkin/zipkin-collector:1.17.0
  environment:
    - MYSQL_HOST=YOUR_HOST
    - MYSQL_USER=YOUR_USER
    - MYSQL_PASS=YOUR_PASS
  entrypoint: /bin/sh
  command: -c "java -jar /zipkin/zipkin-collector.jar -f /collector-mysql.scala"
  expose:
    - 9410
  ports:
    - 9410:9410
```

This instruction applies to `zipkin-web`. As long as that still uses
scribe, all it needs to do is be linked with a collector that talks to
the same database as `zipkin-java-server`.