Add support for HTTP POST /api/v1/spans
With no content type (or application/json) we accept JSON, and with
content-type=application/x-thrift we accept Thrift. In both cases
a List<Span> is POSted to /api/v1/spans. The endpoint returns
immediately with a 202 (ACCEPTED) and the spans are written to the
store in an @Async service method.