Adds support for HTTP POST /api/v1/spans as an alternative to Scribe
It has come up regularly that people are sending spans directly to the
collector's scribe interface, not because they use scribe, but because
they want a simpler architecture when getting started with zipkin. For
example, they are not yet ready for a data pipeline and just want to run
without any dependencies.

While this works, scribe uses Thrift RPC, which is rarely supported, and
for example doesn't have load balancer support (except at TCP level) or
reverse proxies like HTTP does. The type of users who connect directly
to the Collector's scribe endpoint would be better off with an HTTP
option, this is that.

Details of the HTTP Api:

With no content type (or `application/json`) we accept JSON, and with
`content-type=application/x-thrift` we accept Thrift. In both cases
a List is POSTed to `/api/v1/spans` on the query service. The endpoint
returns immediately with a 202 or 400 if the input was malformed.

In either case, spans are written asynchronously to the span store.

Ex. from https://github.com/adrianco/spigo
```bash
$ curl -s localhost:9411/api/v1/spans -X POST --data @json_metrics/cassandra_flow.json -H "Content-Type: application/json"
```