Eliminates case class tech debt in Span and clarifies debug is an option
Before, the common Span type was a strange trait, only used as a case
class. It mismatched the thrift Span struct as it treated debug as if it
weren't an option. This cleans up the tech debt and removes all passing
of `Nil` wrt instantiating a Span class. `Nil` should be very uncommon
in scala code.