Unzips all jar and removes actuator, reducing bootstrap time
With a default docker-machine on a Mid-2014 Macbook Pro, and the default
docker-compose file. I was getting consistent 16s startup, as reported
by Spring Boot's output.

My suspicion was that this was packaging related, as there are many jars
inside the default assembly of Spring Boot. Since Boot does classpath
scanning, it would need to unpack these in order to progress.

@dsyer found that if we do the unpacking at install time, we can reduce
the impact. I verified this, by running the same operation. I'm now
getting consistent 14s ranges on the same docker-machine.

There's other optimizations left. For example, if you remove the
database property and delete the Hikari dep from the pom (so that boot
can't find it), another 2-3 seconds are saved in an otherwise identical
docker-machine.

Regardless, I'm glad to have a couple seconds off without changing
functionality. This change does that.