Fixes service name resolution of dependency graph
Before this change, the dependency graph would display the *first*
serviceName in a span, and use that as the parent name. The problem is
that there are multiple endpoints in a span.

For example, the endpoint associated with "sr" or "cs" is the local
server, in this case the server *or* the client. So, just picking any
endpoint and using it as the service name will certainly be inaccurate.

Moreover, sometimes the server isn't instrumented, so in this case it
won't be able to log an "sr" annotation. Unless you look at the "sa" or
Service Addr annotation, you won't be able to name that span, either.

This fixes both of these problems, by considering only service names of
"sr" or "sa" when building dependency links.

This also backfills tests for the previous span name parsing behavior.