Settles on endTs and lookback query parameters
This settles on endTs as the only time boundary for queries, with a
configurable lookback of either the flag `zipkin.queryService.lookback`
or when running zipkin-query-service, the env variable `QUERY_LOOKBACK`.

We have two queries that need to support a notion of endTs and lookback:
trace and dependency. By supporting lookback, we allow administrators
control over how much data might be subject to search by default.

In the past, the dependency query documentation said it defaulted to
lookback 24hrs. However, the query passed actually applied to all data,
as 0 was passed as the startTs parameter. This changes the default to
be the same as trace query, 7 days, and overridable.

7 days was chosen as that's the length of the data TTL. We would expect
administrators to override this, for example to 2 days.