Drops untested H2, PostgreSQL drivers; switches MySQL to MariaDB
Zipkin formerly pulled dependencies for H2 and PostgreSQL, eventhough we
don't document, test, or allow their configuration in docker or shell
scripts. This drops H2 and PostgreSQL until such time as there's enough
interest to formally support them.

This also switches the MySQL driver to MariaDB. MySQL connector is GPL,
even if it has an OSS exception.
https://www.mysql.com/about/legal/licensing/foss-exception/

MariaDB has a friendlier license, LGPL, which is less scary in audits.