Adds bootstrap and codec to self-tracing
The scala project logs a local trace of bootstrap, to serve as an
example of how to do that. This adds a local trace, including the same
start signal (init), but spring-boot specific annotations for lifecycle
events.

This also adds a SpanStore tracer (except for writes), to show an
example of how trace a local component.