Stops setting Span.timestamp when there's only one annotation
`ApplyTimestampAndDuration` is a bridge that sets Span.timestamp and
duration until instrumentation can be updated to set that directly.

This implies that this feature is only for legacy spans. We know legacy
spans use zipkin core annotations, which come in pairs. For example,
the smallest legacy span is the root span, which has two annotations:
"sr" and "ss".

We found in practice that `ApplyTimestampAndDuration` set Span.timestamp
to the wrong value, the "ss" of a span. This interferes with clock skew
adjustment logic. `ApplyTimestampAndDuration` should only affect legacy
spans, and most specifically complete ones. This changes the code to
abstain from affecting spans with only one annotation, which avoids the
possibility of accidentally choosing "ss" for Span.timestamp.