Adds BeforeTheFactSamplingBenchmarks to support future samplers
Zipkin v1 uses statistical sampling, both locally in instrumentation and
at the collection-tier. For the zipkin-server to be compatible with the
scala collector, we need to implement the same sampling approach.
Specifically, this is to support the `COLLECTOR_SAMPLE_RATE` property.

A follow-up change will implement a collector, using the
`compareTraceId_mostNegativeNumberDefense` measured here. Other
algorithms are measured for the sake of comparison.