Reverts Kafka client to 0.8.x to avoid broker compatibility issues
As it turns out, the upgrade path of kafka is brokers first, as 0.9
brokers are compatible with 0.8 clients, but not visa-versa. The eager
upgrade I made broke that path.

This materialized as an IllegalArgumentException with no message, so
almost impossible to debug.

This change sticks with 0.8.x clients until we either make a standalone
0.9 kafka collector or re-evaluate when 0.9 adoption is higher.

See https://groups.google.com/d/msg/confluent-platform/r8xox149Vn4/qR9q1QjwBQAJ