Removes null coersion in Codec in favor of IllegalArgumentException
When a codec is reading, it is from a byte array, so the only possiblity
of problems is malformed input. This uses IllegalArgumentException to
indicate something the user can fix.

Since all domain objects are locally defined in core, and validated at
construction time, any errors writing to a byte array are bugs. This
uses AssertionError to indicate a bug.

Before, we weren't sending an error message to the user. This sends a
single-line description back in the 400 response. It also moves logging
to the call-site (`ZipkinQueryApiV1`).

Fixes #66