Build frontend assets with Webpack
This lets us use features like ES2015 syntax, commonjs modules,
and makes it easier for developers, with tools like hot-reload.

Bower is replaced by npm, for all libraries except dagre-d3.
We must upgrade to the newest version of dagre-d3 before we can
import it fron npm, but there are breaking changes we need to
account for.

The new tooling requires node/npm on the developer machine, but
this is not a requirement for building zipkin-web. Gradle will
take care of downloading a local copy of node and npm that is
used during the build step.