Uses Anorm in CassandraDependencyStoreSpec
This removes `Dependencies.toLinks` in favor of re-using Anorm for
Cassandra dependency link tests.

Previously, we had a partially implemented aggregator:
`Dependencies.toLinks`. This was added only to make it possible to test
`CassandraDependencyStore`, which doesn't include an aggregator (ex.
zipkin-dependencies-spark is not in this repo).

`Dependencies.toLinks` hasn't been updated for the various changes in
tests and tracking them here is not valuable maintenance. Since the only
purpose was to feed `CassandraDependencyStore`, we can instead use anorm
which is becoming better maintained.