Prefers client when deriving duration
Previously, the clock skew corrector incorrectly derived duration. This
was inaccurate as the distance between annotations on the span's host is
not subject to skew.

When zipkin spans cross multiple hosts, it is usually due to a client+
server span. By preferring the client's duration, we avoid having to
consider skew when deriving duration from annotations. Regardless, we
only derive duration when it isn't explicitly set.