Frontend routing
Introduce crossroads, a JavaScript URL router. This will let us
perform logic client-side based on which page is served. The first
use-case is to only mount the relevant flight.js components for
which page we are looking at.

query-string is an npm module that parses query strings. We
replace the home-baked jquery plugin with this npm module instead.

main.js is simplified and rewritten to ES6.