Move all mustache rendering to the browser
This means the zipkin-web backend is now only
basically a JSON api + delivering static assets.
There's a lot of complexity going on in the
transformation from trace/span JSON data to
data that is fed into mustache templates. These
transformations are still written in Scala, but
hosted at /modelview endpoints. Ideally, they
should be ported to Scala.

Some code had to be moved around - event handlers
must be bound after the HTML is rendered, which now
happens at a later stage than it used to (when it
was server-side rendered).