Fix UI bug - trace placeholder wasn't shown
After the UI refactoring, a bug in the handling of
queryResults surfaced. This means the "placeholder" that is
shown before you have made any queries was never rendered.

This fix adds a "queryWasPerformed" property that is
computed client-side. The structure of the JSON returned
from the server is also simplified a bit.