Adds /health endpoint to zipkin-query and zipkin-web
Various platforms have admin instances that expose health, and Twitter's
does so on a different port. By introducing a "/health" endpoint, we can
start a portable conversation about health that doesn't rely on platform
conventions or opening more ports on the firewall. The use case for this
is load balancer health checks and the zipkin javascript UI.

See #994