Removes internal use of libthrift
This implements TBinaryProtocol directly in the ThriftCodec. By doing
the code, dependencies, and exceptions are simpler (as we are decoding
via a buffer not a network). This uses Okio Buffer, which also
simplified some code.

The payment was having to implement a couple utilities from libthrift,
namely the skip logic. Overall, the code is smaller and the binary
shrunk from 275k to 225k.