Move all UI business logic to the client side
The UI used to leverage a /modelview endpoint on the Scala server
to generate models to pass into its Mustache templates. The logic
from these /modelview endpoint is now ported to JavaScript,
so that all the pre-rendering business logic happens on the
client side.

Since this is a high-risk change, the /modelview endpoints
still exist on the server. Every time we request a page,
the UI will actually fetch a processed version from that
endpoint as well, and compare the Scala version with
the JavaScript version. The idea is that we can compare
outputs of the two, and if there's a difference, we can
assume that's a bug in the JavaScript port.
The endpoints, and the UI logic to compare implementations,
will soon be removed (after testing).