Backfills test that guarantees Span.timestamp, duration precedence
There's logic in ApplyTimestampAndDuration which establishes precedence
of Span.timestamp and duration in shared RPC spans. It is important for
this to be a top-level test as instrumentation need to know what the
rules are. This test makes it explicit that the client wins in shared
spans.

See https://github.com/spring-cloud/spring-cloud-sleuth/issues/188