Adds codec methods needed for Cassandra compatibility
This adds thrift read methods that accept ByteBuffers. This is in
support of the underlying DataStax Cassandra driver, which returns data
in ByteBuffers. The implementation takes care not to re-buffer input
data.

This also introduces an internal type, Dependencies, which is only used
in cassandra: https://github.com/openzipkin/zipkin/issues/1008

See #89