Client-side rendering of traces
This is a JavaScript port of all the Scala business logic for
rendering a trace with Mustache. Now, everything happens
client-side, based on API-calls to zipkin-query.

The same JS-vs-Scala comparison functionality as the one from
the default page is implemented here, to try and detect
regressions by comparing real-life data. This comparison
will be removed before merging to the master branch.

mkDurationStr was updated with a few minor fixes to make it
resemble the original more closely.
shallow-deep-almost-equal is used for the comparison, so that
it assumes 1.0000001 == 1.000000 (To take into account
JavaScript rounding errors)