Changes default cassandra load balancing policy to dc-aware round-robin
Before, the default cassandra load balancing policy was to consider all
datacenters local. This could lead to high latency when storing or
receiving spans. This changes the default policy to be datacenter aware.

The impact is that the first host's datacenter will be chosen as the
local one, and any hosts that aren't in that datacenter will be ignored.

As before, the local datacenter can be pinned via `CASSANDRA_LOCAL_DC`.