Remove aliases from webpack (flight -> flightjs, chosen)
The npm package for flight is actually called 'flightjs'.
Chosen is actually at 'chosen-npm/public/chosen.jquery.js'.
During migration/rewrite, it was convenient to add webpack
aliases, so we wouldn't need to touch the JavaScript code.

Removing the aliases and referring to the packages directly
will reduce technical debt in the long term.