Supports viewing raw traces via GET /api/v1/trace/:id?raw
When troubleshooting instrumentation, we sometimes need to hop to the
database to see what the collector stored (which is closer to what the
instrumentation sent).

In some cases, we are hunting down "clock skew" problems, and sometimes
that's a red herring. Some times we are looking to see if applications
are sending the same span multiple times. The reason we have to go to
the database is we can't quite tell if query-time adjustment logic is
helping or hurting.

This introduces a new query flag "raw" to the get trace endpoint. When
specified, no query time adjustments are made after data is returned
from the backend.