Adds SpanConsumer and settles on List, not Iterator
Before this change, we accepted Iterator of spans instead of a list,
mainly to accomodate the sampling use case. It turns out that this is
awkward, and it also interferes with being able to know the count of
the input. Finally, in most cases the call to accept spans comes from
a codec, and that codec has already decoded a List. This changes the
primary input to a List instead of an Iterator.

This also moves the code that stores spans received by a transport into
a type `SpanConsumer`. This helps reduce the coupling around facets such
as collection-tier sampling, and also makes it easier to write other
transports such as Kafka.