[zipkin-ui] Change nomenclature of tasks calling npm
The current naming implies knowledge of the implementation of the
npm tasks it's calling. I expect in the far future we'll replace
karma with another test framework, and forget to change the name of
the task, leading to confusion and inconsistency.

(Even so there's still some coupling between webpack and Gradle in the
name of the "dist" dir, but if the npm build output dir changes, this
will automatically break. If it doesn't need an update, this will be
just fine)