Serve from /static/ following #1035
This change also revamps the way resources are resolved.

The old logic was "find the static resource, and if you can't, then
serve index.html".

The new logic is "if the request is a file (through heuristic 'it has an
extension'), serve it. otherwise serve index.html"

This makes the intention cleaner, and gives us 404 where a 404 is needed
instead of the contents of index.html