Allows transport-kafka to consume JSON encoded spans
By allowing JSON encoded span lists as Kafka message values, we enable
users to send the same payload as they can over the http transport. It
also allows a very easy debug story, such as sending test spans via CLI.

This implementation peeks at the first byte in the Kafka message, and
attempts to decode json, if it is a '[', or decimal 91. When using
TBinaryProtocol (thrift) encoding, we know that the first byte will be
the Type, which is decimal <=16, so there's no risk of clash.

See https://github.com/openzipkin/zipkin/pull/1052