Simplifies wiring to not attempt to detect classes
`@ConditionalOnClass` has been unreliable in docker, regardless of
whether we unpack the all-jar or not. For example, mysql currently fails
in docker eventhough all of its classes are in the classpath, while
Cassandra works.

We already have a test to ensure optional dependencies are optional,
and we already have a property that is only set when someone wants a
specific storage component (`zipkin.storage.type`). The extra class
checking is unnecessary complexity, so let's not do it.