Unwinds inheritance of InMemorySpanStore
Particularly when using Spring, multiple inheritance can confuse things.
For example, if InMemorySpanStore implements AsyncSpanConsumer, and
another bean does, too, the user will have to explicitly choose which
they want. This complexity isn't helpful.

This unwinds `InMemorySpanStore` to only implement `SpanStore` directly.
In the near future, we will move to a component model which should
reduce this type of concern.