Splits instrumentation sampling from collector sampling
Instrumentation samplers are before-the-fact. When a trace id is
provisioned, a sampler is applied to it to see if tracing should take
place or not. The decision is made once and propagated downstream.

Collector samplers are after-the-fact. When parts of a trace are
received from multiple reporters, they all need to be retained or
dropped consistently.

Before, we were using the same sampler for both sides. This is a
problem in two ways, but mostly for low-traffic sites.

* Low-traffic entry-points will see an unstable sample rate, at worst
dropping all traces. For example, the default RNG in java applied to
the boundary sampler would only get within 4% of the target rate after
100k decisions. Exact precision could be prefered, especially since the
overhead is still less than a microsecond/decision.

* High-traffic entry-points had low-sampling overhead, but they had a
chance of sampling exactly the same trace ids. This would cause two
unrelated requests to show up in the same trace.

This corrects the situation by not sharing a sampling algorithm between
the collectors and the instrumentation. It introduces a counting sampler
which has 100% precision in groups of 100, and a boundary sampler, which
is an order of magnitude faster, but probablistic with a measured 4%
error rate after 100K samples.